requirements.txt
streamlit
transformers
torch
huggingface_hub
python-dotenv
